Rollback version changes and Store SSO Token even when SSO is not configured
