+++ app/src/main/res/values/strings.xml
@@ -16,14 +16,14 @@
     <!--  Forgerock SDK Configuration -->
 
     <!-- OAuth -->
-    <string name="forgerock_oauth_client_id" translatable="false">andy_app2</string>
-    <string name="forgerock_oauth_redirect_uri" translatable="false">https://www.example2.com/callback</string>
+    <string name="forgerock_oauth_client_id" translatable="false">andy_app</string>
+    <string name="forgerock_oauth_redirect_uri" translatable="false">http://www.example2.com/callback</string>
     <string name="forgerock_oauth_scope" translatable="false">openid profile email address phone</string>
-    <string name="forgerock_oauth_url" translatable="false">https://openam.example.com:8081/openam</string>
+    <string name="forgerock_oauth_url" translatable="false">http://openam.example.com:8081/openam</string>
     <integer name="forgerock_oauth_threshold" translatable="false">30</integer> <!-- in second -->
 
     <!-- Server -->
-    <string name="forgerock_url" translatable="false">https://openam.example.com:8081/openam</string>
+    <string name="forgerock_url" translatable="false">http://openam.example.com:8081/openam</string>
     <string name="forgerock_realm" translatable="false">root</string>
     <integer name="forgerock_timeout" translatable="false">30</integer>
